backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==10.1.0
opencv-python==4.8.1.78
numpy>=1.26.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
openai>=1.0.0
requests>=2.31.0
boto3>=1.34.0
